requirements.txt
pandas
seaborn
matplotlib.pyplot
tabulate
scikit-learn
xgboost
lazypredict
imbalanced-learn
yellowbrick
itertools